新增菜单显示功能

+++ src/main/java/com/ruoyi/framework/core/controller/BaseController.java
@@ -15,7 +15,7 @@ public class BaseController
         return ShiroUtils.getUser();
     }
 
-    public Integer getUserId()
+    public Long getUserId()
     {
         return getUser().getUserId();
     }
